requirements.txt
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
Django==5.0.1
django-cleanup==8.0.0
djangorestframework==3.14.0
fairscale==0.4.13
filelock==3.13.1
fsspec==2023.12.2
google-auth==2.26.0
google-auth-oauthlib==1.2.0
grpcio==1.60.0
grpcio-tools==1.60.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.20.2
hyperframe==6.0.1
idna==3.6
Jinja2==3.1.2
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
oauthlib==3.2.2
packaging==23.2
pillow==10.2.0
portalocker==2.8.2
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.3
pydantic_core==2.14.6
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
qdrant-client==1.7.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.1
sniffio==1.3.0
sqlparse==0.4.4
sympy==1.12
timm==0.9.12
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==1.26.18
